+++ sdk/open_api/configuration.py
@@ -497,7 +497,7 @@ class Configuration:
                "OS: {env}\n"\
                "Python Version: {pyversion}\n"\
                "Version of the API: 3.0.14\n"\
-               "SDK Package Version: 3.0.14.0".\
+               "SDK Package Version: 3.0.15.0".\
                format(env=sys.platform, pyversion=sys.version)
 
     def get_host_settings(self) -> List[HostSetting]: